+++ scim2_client/engines/httpx2.py
@@ -143,7 +143,7 @@ class SyncSCIMClient(BaseSyncSCIMClient):
 
     def query(
         self,
-        resource_model: type[Resource] | None = None,
+        target: type[Resource] | Resource | None = None,
         id: str | None = None,
         query_parameters: ResponseParameters | dict | None = None,
         check_request_payload: bool | None = None,
@@ -158,7 +158,7 @@ class SyncSCIMClient(BaseSyncSCIMClient):
             query_parameters, search_request
         )
         req = self._prepare_query_request(
-            resource_model=resource_model,
+            target=target,
             id=id,
             query_parameters=query_parameters,
             check_request_payload=check_request_payload,
@@ -217,8 +217,8 @@ class SyncSCIMClient(BaseSyncSCIMClient):
 
     def delete(
         self,
-        resource_model: type[Resource],
-        id: str,
+        resource: Resource | type[Resource] | None = None,
+        id: str | None = None,
         check_response_payload: bool | None = None,
         expected_status_codes: list[int]
         | None = BaseSyncSCIMClient.DELETION_RESPONSE_STATUS_CODES,
@@ -226,7 +226,7 @@ class SyncSCIMClient(BaseSyncSCIMClient):
         **kwargs,
     ) -> Error | dict | None:
         req = self._prepare_delete_request(
-            resource_model=resource_model,
+            resource=resource,
             id=id,
             expected_status_codes=expected_status_codes,
             **kwargs,
@@ -279,9 +279,9 @@ class SyncSCIMClient(BaseSyncSCIMClient):
 
     def modify(
         self,
-        resource_model: type[ResourceT],
-        id: str,
-        patch_op: PatchOp[ResourceT] | dict,
+        resource: ResourceT | type[ResourceT] | None = None,
+        patch_op: PatchOp[ResourceT] | dict | None = None,
+        id: str | None = None,
         check_request_payload: bool | None = None,
         check_response_payload: bool | None = None,
         expected_status_codes: list[int]
@@ -290,9 +290,9 @@ class SyncSCIMClient(BaseSyncSCIMClient):
         **kwargs,
     ) -> ResourceT | Error | dict | None:
         req = self._prepare_patch_request(
-            resource_model=resource_model,
-            id=id,
+            resource=resource,
             patch_op=patch_op,
+            id=id,
             check_request_payload=check_request_payload,
             expected_status_codes=expected_status_codes,
             **kwargs,
@@ -374,7 +374,7 @@ class AsyncSCIMClient(BaseAsyncSCIMClient):
 
     async def query(
         self,
-        resource_model: type[Resource] | None = None,
+        target: type[Resource] | Resource | None = None,
         id: str | None = None,
         query_parameters: ResponseParameters | dict | None = None,
         check_request_payload: bool | None = None,
@@ -389,7 +389,7 @@ class AsyncSCIMClient(BaseAsyncSCIMClient):
             query_parameters, search_request
         )
         req = self._prepare_query_request(
-            resource_model=resource_model,
+            target=target,
             id=id,
             query_parameters=query_parameters,
             check_request_payload=check_request_payload,
@@ -450,8 +450,8 @@ class AsyncSCIMClient(BaseAsyncSCIMClient):
 
     async def delete(
         self,
-        resource_model: type[Resource],
-        id: str,
+        resource: Resource | type[Resource] | None = None,
+        id: str | None = None,
         check_response_payload: bool | None = None,
         expected_status_codes: list[int]
         | None = BaseAsyncSCIMClient.DELETION_RESPONSE_STATUS_CODES,
@@ -459,7 +459,7 @@ class AsyncSCIMClient(BaseAsyncSCIMClient):
         **kwargs,
     ) -> Error | dict | None:
         req = self._prepare_delete_request(
-            resource_model=resource_model,
+            resource=resource,
             id=id,
             expected_status_codes=expected_status_codes,
             **kwargs,
@@ -514,9 +514,9 @@ class AsyncSCIMClient(BaseAsyncSCIMClient):
 
     async def modify(
         self,
-        resource_model: type[ResourceT],
-        id: str,
-        patch_op: PatchOp[ResourceT] | dict,
+        resource: ResourceT | type[ResourceT] | None = None,
+        patch_op: PatchOp[ResourceT] | dict | None = None,
+        id: str | None = None,
         check_request_payload: bool | None = None,
         check_response_payload: bool | None = None,
         expected_status_codes: list[int]
@@ -525,9 +525,9 @@ class AsyncSCIMClient(BaseAsyncSCIMClient):
         **kwargs,
     ) -> ResourceT | Error | dict | None:
         req = self._prepare_patch_request(
-            resource_model=resource_model,
-            id=id,
+            resource=resource,
             patch_op=patch_op,
+            id=id,
             check_request_payload=check_request_payload,
             expected_status_codes=expected_status_codes,
             **kwargs,